Remove client integration in favor of client symbol provider class
