--- codegen/client/src/main/resources/META-INF/services/software.amazon.smithy.java.codegen.JavaCodegenIntegration
@@ -1,1 +1,0 @@
-software.amazon.smithy.java.codegen.client.integration.ClientIntegration
